Bug 1088216 - docs: make the installation instructions easier to follow
This separates out the steps required for running the tests, from those
for setting up a local instance. The running ingestion tasks step now
clearly states the need for the API server to be running already. The
log parser compile step is moved inline since it was otherwise easy to
omit.